refactor funcs
move funcs related to parsing CAZy data to the new  module. The  module contains only funcs related to crawling the CAZy webiste and making calls to the CAZy website
