+++ scraper/cazy_webscraper.py
@@ -80,6 +80,7 @@ from typing import List, Optional
 from tqdm import tqdm
 
 from scraper import crawler
+from scraper import cazy
 from scraper.sql import sql_orm, sql_interface
 from scraper.utilities import (
     build_logger,
@@ -366,15 +367,12 @@ def get_cazy_data(
         sys.exit
     
     # extract the CAZy family data and add to the local CAZyme database
-    cazy_txt_lines = crawler.extract_cazy_file_data(cazy_txt_path)
+    cazy_txt_lines = cazy.extract_cazy_file_data(cazy_txt_path)
     logger.info(f"Retrieved {len(cazy_txt_lines)} lines from the CAZy db txt file")
 
     (
         cazy_data,
-        families_db_insert_values,
-        kingdoms_db_insert_values,
-        taxonomy_db_insert_values,
-    ) = crawler.parse_cazy_data(
+    ) = cazy.parse_cazy_data(
         cazy_txt_lines,
         class_filters,
         fam_filters,
@@ -383,19 +381,11 @@ def get_cazy_data(
         cazy_fam_populations,
     )
     logger.info(
-        f"Retrieved f{len((list(cazy_data.keys())))} proteins from the CAZy txt file "
+        f"Retrieved {len((list(cazy_data.keys())))} proteins from the CAZy txt file "
         "matching the scraping criteria"
     )
 
-    genbank_db_insert_values = [(gbk_accession,) for gbk_accession in (list(cazy_data.keys()))]
-
-    # Insert values into the local CAZyme database
-    sql_interface.insert_data(connection, 'Genbanks', ['genbank_accession'], genbank_db_insert_values)
-    sql_interface.insert_data(connection, 'CazyFamilies', ['family', 'subfamily'], families_db_insert_values)
-    sql_interface.insert_data(connection, 'Taxs', ['genus', 'species'], taxonomy_db_insert_values)
-    sql_interface.insert_data(connection, 'Kingdoms', ['kingdom'], kingdoms_db_insert_values)
-
-    # update relationships
+    ####
 
     return
